// frontend/src/types/OverwriteServerRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type OverwriteServerRequest = { server_id: number, port: number | null, version_name: string | null, protocol: number | null, is_online_mode: boolean | null, is_forge: boolean | null, is_online: boolean | null, ping: bigint | null, favicon: string | null, is_checked: boolean | null, is_spoofable: boolean | null, is_crashed: boolean | null, };